requirements.txt
pydantic
python-dotenv
sendgrid
ipython
typing-extensions
ipykernel
openai-agents
resend
gradio
streamlit